refactor(json): extract readJson to src/json.ts
