+++ src/mcp-config.ts
@@ -1,5 +1,5 @@
+import { readJson } from "./json";
 import { type McpConfig, mcpConfigSchema } from "./mcp-contract";
-import { readJson } from "./workspace-detectors";
 
 export function readMcpConfig(workspace: string): McpConfig {
   const raw = readJson(workspace, ".mcp.json");
