+++ src/workspace-detectors.ts
@@ -1,24 +1,10 @@
 import { existsSync, readFileSync } from "node:fs";
 import { join } from "node:path";
+import { readJson } from "./json";
 import type { WorkspaceCommand, WorkspaceProfile } from "./workspace-profile";
 
 export function fileExists(workspace: string, name: string): boolean {
   return existsSync(join(workspace, name));
-}
-
-function stripJsonComments(text: string): string {
-  return text.replace(/\/\/.*$/gm, "").replace(/\/\*[\s\S]*?\*\//g, "");
-}
-
-export function readJson(workspace: string, name: string): Record<string, unknown> | null {
-  try {
-    const path = join(workspace, name);
-    if (!existsSync(path)) return null;
-    const raw = readFileSync(path, "utf8");
-    return JSON.parse(name.endsWith("c") ? stripJsonComments(raw) : raw) as Record<string, unknown>;
-  } catch {
-    return null;
-  }
 }
 
 export function readText(workspace: string, name: string): string | null {
